Лабораторна 15/requirements.txt
flask~=3.0.0
wtforms~=3.1.1
alembic~=1.12.1
SQLAlchemy~=2.0.23
Pillow~=10.1.0
app~=0.0.1
werkzeug~=3.0.1
PyJWT~=2.8.0
marshmallow~=3.20.1